+++ src/main/java/de/samply/beam/BeamRequestMetadata.java
@@ -1,14 +1,18 @@
-package de.samply.exporter.focus;
+package de.samply.beam;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 @Data
+@JsonInclude(JsonInclude.Include.NON_NULL)
 public class BeamRequestMetadata {
 
     @JsonProperty("project")
     private String project;
     @JsonProperty("task_type")
     private TaskType taskType;
+    @JsonProperty("transform")
+    private String transform;
 
 }