+++ src/main/java/de/samply/feasibility/FeasibilityServiceException.java
@@ -1,0 +1,12 @@
+package de.samply.feasibility;
+
+public class FeasibilityServiceException extends RuntimeException {
+
+    public FeasibilityServiceException(String message) {
+        super(message);
+    }
+
+    public FeasibilityServiceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}